Reorganize library structure
- move devices modules in devices subfolder
- correct related import
