+++ custom_components/smartthinq_sensors/humidifier.py
@@ -21,7 +21,7 @@ from homeassistant.helpers.update_coordinator import CoordinatorEntity
 from . import LGEDevice
 from .const import DOMAIN, LGE_DEVICES, LGE_DISCOVERY_NEW
 from .wideq import FEAT_HUMIDITY, FEAT_TARGET_HUMIDITY, DeviceType
-from .wideq.dehumidifier import DeHumidifierDevice
+from .wideq.devices.dehumidifier import DeHumidifierDevice
 
 ATTR_CURRENT_HUMIDITY = "current_humidity"
 ATTR_FAN_MODE = "fan_mode"
