+++ custom_components/smartthinq_sensors/wideq/devices/__init__.py
@@ -1,0 +1,1 @@
+"""Support for LG ThinQ devices."""
