+++ custom_components/smartthinq_sensors/wideq/devices/dehumidifier.py
@@ -3,9 +3,9 @@ from enum import Enum
 import logging
 from typing import Optional
 
-from .const import FEAT_HUMIDITY, FEAT_TARGET_HUMIDITY, FEAT_WATER_TANK_FULL
-from .core_exceptions import InvalidRequestError
-from .device import Device, DeviceStatus
+from ..const import FEAT_HUMIDITY, FEAT_TARGET_HUMIDITY, FEAT_WATER_TANK_FULL
+from ..core_exceptions import InvalidRequestError
+from ..device import Device, DeviceStatus
 
 CTRL_BASIC = ["Control", "basicCtrl"]
 STATE_POWER_V1 = "InOutInstantPower"
