+++ custom_components/smartthinq_sensors/wideq/devices/range.py
@@ -2,7 +2,7 @@
 import logging
 from typing import Optional
 
-from .const import (
+from ..const import (
     BIT_OFF,
     FEAT_COOKTOP_CENTER_STATE,
     FEAT_COOKTOP_LEFT_FRONT_STATE,
@@ -18,7 +18,7 @@ from .const import (
     UNIT_TEMP_CELSIUS,
     UNIT_TEMP_FAHRENHEIT,
 )
-from .device import Device, DeviceStatus, UnitTempModes
+from ..device import Device, DeviceStatus, UnitTempModes
 
 OVEN_TEMP_UNIT = {
     "0": UnitTempModes.Fahrenheit,
